refactor(access): Move models to their own file
ref: #640 #641
